Move discovery-cap graph-slice coverage
Move the discovery-cap spillover test into the graph-slice coverage child
module so the parent test module stays below the file-size limit. Add the
`--max-cards 0` invalid-argument case and strengthen spillover frontier
assertions to validate each emitted frontier entry.
